Default max-count to infinite. This fixes #516.

+++ src/options.h
@@ -48,7 +48,7 @@ typedef struct {
     int literal;
     int literal_starts_wordchar;
     int literal_ends_wordchar;
-    int max_matches_per_file;
+    size_t max_matches_per_file;
     int max_search_depth;
     char path_sep;
     char *path_to_agignore;
